+++ configure.ac
@@ -125,6 +125,7 @@ AS_IF([test "x$with_zstd" != "xno"], [
 	AC_MSG_NOTICE([Zstandard support not requested])
 ])
 CC_FEATURE_APPEND([with_features], [with_zstd], [ZSTD])
+AM_CONDITIONAL([ENABLE_ZSTD], [test "x$with_zstd" != "xno"])
 
 AC_ARG_WITH([xz],
 	AS_HELP_STRING([--with-xz], [handle Xz-compressed modules @<:@default=disabled@:>@]),
